docs: update docs for scripts and sass files + update ReadMe.md
