+++ ReadMe.md
@@ -1,6 +1,7 @@
 # Правила проекта
 
 1. Каждый раздел верстки делается в отдельной ветки, которая сливается с основной только после PullRequest.
+   Ответвления делаем от ветки Develop, ветка main существует для релизных версий приложения.
 2. Чужой код (и файлы) трогаем только на CodeRewiew, в присутствии автора.
 3. Стараемся писать независимые разделы кода, чтобы не создавать конфликтов с другими участниками.
 4. По распределению/отслеживанию задач и общению: предлагаю телеграмм, ибо канбан доски сильно муторно.
@@ -44,9 +45,9 @@
 
 # Стиль проекта
 
-- Именование файлов - с маленькой буквы, кебаб нотация (исключения: js классы, )
+- Именование файлов - с маленькой буквы, кебаб нотация (исключения: js классы)
 - БЭМ блок = отдельному файлу стилей
--
+- Делить большие файлы на модули - это хорошо
 
 # Указываем тип коммита
 
@@ -56,7 +57,7 @@
 - style — исправляем опечатки, исправляем форматирование
 - refactor — рефакторинг кода приложения
 
-**Лучше написать: _добавил футер_, чем просто: _футер_. Что ты с этим футером сделал? Удалил, добавил, изменил, проклял? **
+**Лучше написать: _добавил футер_, чем просто: _футер_. Что ты с этим футером сделал? Удалил, добавил, изменил, проклял?**
 
 # Как заставить sass компилироваться в папку styles/css ?
 
@@ -90,7 +91,7 @@
 
 - Или установить Prettier форматтером по умолчанию: configute default formatter (на прошлом скриншоте)
 
-# Как начать внести вклад в проект если ты не коллаборатор?
+### Как начать внести вклад в проект если ты не коллаборатор?
 
 1. Делаем форк данного репозитория
