+++ pages/main.html
@@ -9,12 +9,16 @@
     />
     <link rel="stylesheet" href="../styles/css/main.css" />
     <script src="../lib/anime.min.js" defer></script>
-    <script src="../scripts/main.js" defer></script>
+    <script
+      src="../scripts/main.js"
+      type="module"
+      defer
+    ></script>
     <title>VRNas</title>
   </head>
 
   <body>
-    <svg>
+    <svg class="visually-hidden">
       <defs>
         <linearGradient
           id="paint0_linear_28530_2192"
