+++ scripts/check-icons-animation.js
@@ -1,0 +1,31 @@
+/**
+ * @module - модуль описывающий работу анимации морфинга иконок в секции about-us
+ *
+ * Иконки поочередно переходят из неактивного состояния в активное путем svg морфинга (переход из одной svg в другую, если грубо говорить)
+ */
+
+/**
+ * @constant Точки, по которым осуществляется морфинг
+ */
+const svgPaths = {
+  d0: 'M14.2222 14.2222H1.77778V1.77778H10.6667V0H1.77778C0.8 0 0 0.8 0 1.77778V14.2222C0 15.2 0.8 16 1.77778 16H14.2222C15.2 16 16 15.2 16 14.2222V7.11111H14.2222V14.2222Z',
+  d1: 'M4.35556 6.31111L3.11111 7.55556L7.11111 11.5556L16 2.66667L14.7556 1.42222L7.11111 9.06667L4.35556 6.31111ZM14.2222 14.2222H1.77778V1.77778H10.6667V0H1.77778C0.8 0 0 0.8 0 1.77778V14.2222C0 15.2 0.8 16 1.77778 16H14.2222C15.2 16 16 15.2 16 14.2222V7.11111H14.2222V14.2222Z',
+}
+
+/**
+ * Описание анимации
+ */
+const checkIconConfig = {
+  targets: '.check-icon path',
+  d: [svgPaths.d0, svgPaths.d1],
+  easing: 'easeInOutExpo',
+  duration: 1000,
+  delay: anime.stagger(300),
+}
+
+/**
+ * Функция инициализатор анимации
+ */
+export function initCheckIcon() {
+  anime(checkIconConfig)
+}
