+++ styles/sass/globals/_base.scss
@@ -1,3 +1,5 @@
+/// Модуль предназначенный для определения глобальных стилей приложения.
+
 @use '../helpers' as *;
 
 :root {
@@ -51,6 +53,7 @@ h6 {
   text-align: center;
 }
 
+/// Класс для градиентных заголовков, обычно дублирующих название секции
 .h2 {
   font-weight: 500;
   letter-spacing: 0.14em;
@@ -61,6 +64,7 @@ h6 {
   -webkit-text-fill-color: transparent;
 }
 
+/// Класс для информационных заголовкой, несущих в себе бизнес ценность
 .h3 {
   font-weight: 600;
   font-size: fluid(48, 32);
