+++ styles/sass/globals/_fonts.scss
@@ -1,1 +1,3 @@
+/// Модуль подключения шрифтов
+
 @import url('https://fonts.googleapis.com/css2?family=Chakra+Petch:wght@300;400;500;600;700&display=swap');
